+++ src/main/java/net/minecraftforge/fmllegacy/server/ServerLifecycleHooks.java
@@ -1,3 +1,8 @@
+/*
+ * Copyright (c) Forge Development LLC and contributors
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+
 package net.minecraftforge.fmllegacy.server;
 
 import net.minecraft.server.MinecraftServer;
@@ -10,6 +15,7 @@ import net.minecraftforge.mixin.accessor.LevelResourceAccessor;
 
 import java.nio.file.Path;
 
+// Forge Config API Port: class greatly reduced to only contain code related to configs
 public class ServerLifecycleHooks {
     private static final LevelResource SERVERCONFIG = LevelResourceAccessor.create(ConfigPaths.SERVER_CONFIG_PATH);
 
@@ -20,10 +26,12 @@ public class ServerLifecycleHooks {
     }
 
     public static void handleServerAboutToStart(final MinecraftServer server) {
+        // Forge Config API Port: removed everything else which wasn't config related
         ConfigTracker.INSTANCE.loadConfigs(ModConfig.Type.SERVER, getServerConfigPath(server));
     }
 
     public static void handleServerStopped(final MinecraftServer server) {
+        // Forge Config API Port: removed everything else which wasn't config related
         ConfigTracker.INSTANCE.unloadConfigs(ModConfig.Type.SERVER, getServerConfigPath(server));
     }
 }