+++ src/main/java/net/minecraftforge/network/client/config/ConfigSyncClient.java
@@ -1,5 +1,11 @@
+/*
+ * Copyright (c) Forge Development LLC and contributors
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+
 package net.minecraftforge.network.client.config;
 
+import com.mojang.logging.LogUtils;
 import net.fabricmc.fabric.api.client.networking.v1.ClientLoginNetworking;
 import net.fabricmc.fabric.api.networking.v1.PacketByteBufs;
 import net.minecraft.client.Minecraft;
@@ -7,11 +13,18 @@ import net.minecraft.network.FriendlyByteBuf;
 import net.minecraftforge.fml.config.ConfigTracker;
 import net.minecraftforge.network.client.NetworkHooks;
 import net.minecraftforge.network.config.ConfigSync;
+import org.slf4j.Logger;
 
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
 
+import static net.minecraftforge.network.config.ConfigSync.FMLHSMARKER;
+
+// Forge Config API Port: a class by that name does not exist on Forge, though this is an extract from net.minecraftforge.network.HandshakeHandler
+// with the client side code regarding configs
 public class ConfigSyncClient {
+    private static final Logger LOGGER = LogUtils.getLogger();
+
     public static final ConfigSyncClient INSTANCE = new ConfigSyncClient(ConfigTracker.INSTANCE);
     private final ConfigTracker tracker;
 
@@ -22,14 +35,14 @@ public class ConfigSyncClient {
     public void clientInit() {
         ClientLoginNetworking.registerGlobalReceiver(ConfigSync.SYNC_CONFIGS_CHANNEL, (client, handler, buf, listenerAdder) -> {
             final String fileName = this.receiveSyncedConfig(buf);
-            ConfigSync.LOGGER.debug(ConfigSync.FMLHSMARKER, "Received config sync for {} from server", fileName);
+            LOGGER.debug(FMLHSMARKER, "Received config sync for {} from server", fileName);
             FriendlyByteBuf response = PacketByteBufs.create();
             response.writeUtf(fileName);
-            ConfigSync.LOGGER.debug(ConfigSync.FMLHSMARKER, "Sent config sync for {} to server", fileName);
+            LOGGER.debug(FMLHSMARKER, "Sent config sync for {} to server", fileName);
             return CompletableFuture.completedFuture(response);
         });
         ClientLoginNetworking.registerGlobalReceiver(ConfigSync.MODDED_CONNECTION_CHANNEL, (client, handler, buf, listenerAdder) -> {
-            ConfigSync.LOGGER.debug(ConfigSync.FMLHSMARKER, "Received modded connection marker from server");
+            LOGGER.debug(FMLHSMARKER, "Received modded connection marker from server");
             NetworkHooks.setModdedConnection();
             return CompletableFuture.completedFuture(PacketByteBufs.create());
         });