+++ src/main/java/net/minecraftforge/server/command/ConfigCommand.java
@@ -1,11 +1,16 @@
+/*
+ * Copyright (c) Forge Development LLC and contributors
+ * SPDX-License-Identifier: LGPL-2.1-only
+ */
+
 package net.minecraftforge.server.command;
 
 import com.mojang.brigadier.CommandDispatcher;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.context.CommandContext;
-import net.minecraft.ChatFormatting;
 import net.minecraft.commands.CommandSourceStack;
 import net.minecraft.commands.Commands;
+import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.ClickEvent;
 import net.minecraft.network.chat.Component;
 import net.minecraftforge.fml.config.ConfigTracker;
@@ -15,12 +20,21 @@ import java.io.File;
 
 public class ConfigCommand {
     public static void register(CommandDispatcher<CommandSourceStack> dispatcher) {
-        dispatcher.register(Commands.literal("config").then(ShowFile.register()));
+        dispatcher.register(
+                Commands.literal("config").
+                        then(ShowFile.register())
+        );
     }
 
     public static class ShowFile {
         static ArgumentBuilder<CommandSourceStack, ?> register() {
-            return Commands.literal("showfile").requires(cs -> cs.hasPermission(0)).then(Commands.argument("mod", ModIdArgument.modIdArgument()).then(Commands.argument("type", EnumArgument.enumArgument(ModConfig.Type.class)).executes(ShowFile::showFile)));
+            return Commands.literal("showfile").
+                    requires(cs->cs.hasPermission(0)).
+                    then(Commands.argument("mod", ModIdArgument.modIdArgument()).
+                            then(Commands.argument("type", EnumArgument.enumArgument(ModConfig.Type.class)).
+                            executes(ShowFile::showFile)
+                    )
+            );
         }
 
         private static int showFile(final CommandContext<CommandSourceStack> context) {
@@ -29,9 +43,14 @@ public class ConfigCommand {
             final String configFileName = ConfigTracker.INSTANCE.getConfigFileName(modId, type);
             if (configFileName != null) {
                 File f = new File(configFileName);
-                context.getSource().sendSuccess(Component.translatable("commands.config.getwithtype", modId, type, Component.literal(f.getName()).withStyle(ChatFormatting.UNDERLINE).withStyle((style) -> style.withClickEvent(new ClickEvent(ClickEvent.Action.OPEN_FILE, f.getAbsolutePath())))), true);
+                context.getSource().sendSuccess(Component.translatable("commands.config.getwithtype",
+                        modId, type,
+                        Component.literal(f.getName()).withStyle(ChatFormatting.UNDERLINE).
+                                withStyle((style) -> style.withClickEvent(new ClickEvent(ClickEvent.Action.OPEN_FILE, f.getAbsolutePath())))
+                ), true);
             } else {
-                context.getSource().sendSuccess(Component.translatable("commands.config.noconfig", modId, type), true);
+                context.getSource().sendSuccess(Component.translatable("commands.config.noconfig", modId, type),
+                        true);
             }
             return 0;
         }