+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v4.2.4-1.19.2] - 2022-08-30
+- Resolved issues regarding the license of Forge
+### Changed
+- Ported more changes from Forge's config screens
+- Switched remaining loggers from Log4j to SLF4J
+
 ## [v4.2.3-1.19.2] - 2022-08-22
 ### Fixed
 - Fixed config screen crashing for config values without a widget factory